+++ triplea/client/topic_extraction.py
@@ -80,6 +80,7 @@ def extract_topic(text: str, method: str, top: int = 10, threshold: float = 0) -
     # URL = f"{SETTINGS.AAA_TOPIC_EXTRACT_ENDPOINT}/topic/"
     URL = f"{SETTINGS.AAA_TOPIC_EXTRACT_ENDPOINT}/"
 
+
     # data to be sent to api
     data = {
         "Text": text.replace("\n", " "),